boot_requirements_checher_test broken in smaller pieces
